requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
django-environ>=0.11,<1.0
twilio>=9.0,<10.0
django-apscheduler>=0.7,<1.0
django-cors-headers>=4.4,<5.0
